Updated with group quota

+++ prereq.sh
@@ -20,12 +20,12 @@ apt install uidmap -y
 
 # Check for pattern <Not starting with a comment><Not whitespace(Device)><Whitespace></><Whitespace><Not whitespace(FS type)><Whitespace><No whitespace(Options)><Whitespace><Number(Dump)><Whitespace><Number(Pass)>
 # And whether Options is <Not whitespace>*usrquota<Not whitespace>*
-# If not add usrquota to the options.
+# If not add usrquota and groupquota to the options.
 updated=0
-sed -n -r -e "/^[^#]\S+\s+\/\s+\S+\s+\S+\s+[0-9]+\s+[0-9]+\s*/{ /^\S+\s+\/\s+\S+\s+\S*usrjquota=aquota.user,jqfmt=vfsv0\S*/{q100} }" "$tmpfstab"
+sed -n -r -e "/^[^#]\S+\s+\/\s+\S+\s+\S+\s+[0-9]+\s+[0-9]+\s*/{ /^\S+\s+\/\s+\S+\s+\S*usrjquota=aquota.user,grpjquota=aquota.group,jqfmt=vfsv0\S*/{q100} }" "$tmpfstab"
 res=$?
 if [ $res -eq 0 ]; then
-    sed -i -r -e "/^[^#]\S+\s+\/\s+\S+\s+\S+\s+[0-9]+\s+[0-9]+\s*/{ s/^\S+\s+\/\s+\S+\s+\S+/&,usrjquota=aquota.user,jqfmt=vfsv0/ }" "$tmpfstab"
+    sed -i -r -e "/^[^#]\S+\s+\/\s+\S+\s+\S+\s+[0-9]+\s+[0-9]+\s*/{ s/^\S+\s+\/\s+\S+\s+\S+/&,usrjquota=aquota.user,grpjquota=aquota.group,jqfmt=vfsv0/ }" "$tmpfstab"
     res=$?
     updated=1
 fi
@@ -47,8 +47,8 @@ else
     echo "fstab already configured."
 fi
 
-# Check and turn on user quota if not enabled.
-if [ ! -f /aquota.user ]; then
+# Check and turn on user quota and group quota if not enabled.
+if [ ! -f /aquota.user ] && [ ! -f /aquota.group ]; then
     # quota package is not installed.
     if ! command -v quota &>/dev/null; then
         apt-get install -y quota
